% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_utils.R
\name{apply_theme}
\alias{apply_theme}
\title{applies preset theme options to figures}
\usage{
apply_theme()
}
\value{

}
\description{
Applies preset theme options to figures including fonts, grids, axes
}
\details{
Applies PT Sans font, blank element background, sizes for axis texts,
and other thematic options to keep
}
\author{
Daniel James Liden
}
