% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{druguse}
\alias{druguse}
\title{youth and all-ages drug use metrics in NV and the US}
\format{
a tibble with 48 rows and 4 columns:
\describe{
\item{state:}{Nevada or United States}
\item{metric:}{rate being estimated}
\item{age_group:}{12-17, 18-25, or 26+}
\item{estimate:}{percentage}
}
}
\source{
\url{https://www.samhsa.gov/data/report/2017-2018-nsduh-state-specific-tables}
}
\usage{
druguse
}
\description{
A dataset containing estimates of rates of illicit drug use,
need for treatment for substance abuse, mental illness, and
reception of mental health services.
}
\keyword{datasets}
