% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ed_attainment}
\alias{ed_attainment}
\title{educational attainment by age group in Nevada}
\format{
a data frame with 16 rows and 4 columns:
\describe{
\item{attainment:}{level of educational attainment}
\item{percent:}{percent of Nevadans within age group with specified level of education}
\item{age:}{age group, 18-24 or 25+}
\item{geo:}{geography (Nevada or National)
}
}
\source{
\url{https://data.census.gov/cedsci/table?q=Educational\%20Attainment&g=0400000US32&tid=ACSST5Y2019.S1501&moe=false&hidePreview=true}
}
\usage{
ed_attainment
}
\description{
A dataset showing educational attainment by age group for Nevadans
aged 18-24 and 25+. Education categories are:
\itemize{
\item Less than high school
\item High school graduate or equivalent
\item Some college or associate's degree
\item Bachelor's degree or Higher
}
}
\keyword{datasets}
