% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadownload.R
\name{fetch_pnv}
\alias{fetch_pnv}
\title{}
\usage{
fetch_pnv(path = "~/data_share/youth/")
}
\arguments{
\item{path}{}
}
\value{

}
\description{
get person-level PUMS data
}
\details{
2019 1-year person-level PUMS data with urban/rural classification
Clark and Washoe Counties classified as urban; rest classified as rural
}
\author{
Daniel James Liden
}
