% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_utils.R
\name{fig_base}
\alias{fig_base}
\title{base plot for themed figures}
\usage{
fig_base(data = NULL)
}
\arguments{
\item{data}{data frame or tibble if some default data are used throughout}
}
\value{
ggplot object
}
\description{
Sets up base theme for figures
}
\details{
Sets up base theme for figures by modifying the ggplot classic theme with a
few Guinn Center branded elements
}
\author{
Daniel James Liden
}
