% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druguse.R
\name{get_metrics}
\alias{get_metrics}
\title{}
\usage{
get_metrics(data = druguse)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
unique metrics in drug use dataset
}
\details{
Helper function for obtaining vector of names of unique
metrics in the drug use survey data. Used to help with
plotting.
}
\author{
Daniel James Liden
}
