% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty.R
\name{get_poverty_PUMS}
\alias{get_poverty_PUMS}
\title{Poverty from PUMS}
\usage{
get_poverty_PUMS(
  age_breaks = c(14, 18, 24, 99),
  labels = c("14-18", "19-24", "25+")
)
}
\arguments{
\item{age_breaks}{}

\item{labels}{}
}
\value{

}
\description{
Pull income-to-poverty ratio from PUMS
}
\details{
Poverty status is determined for all people except institutionalized people,
people in military group quarters, people in college dorms, and unrelated
individuals under the age of 15. These were excluded from the analysis.
}
\author{
Daniel James Liden
}
