% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_demo.R
\name{get_pums_race}
\alias{get_pums_race}
\title{}
\usage{
get_pums_race(
  age_breaks = c(0, 14, 24, 99),
  age_labels = c("0-13", "14-24", "25+")
)
}
\arguments{
\item{age_breaks}{min age, max age, and any breaks in between (numerical)}

\item{age_labels}{vector of strings showing age ranges e.g. c("14-24", "25+")}
}
\value{
tibble
}
\description{
race by age group
}
\details{
numbers/percentages of Nevadans by race and age group.
}
\author{
Daniel James Liden
}
