% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_demo.R
\name{get_rural_PUMS}
\alias{get_rural_PUMS}
\title{Age by Urban/Rural County Classification}
\usage{
get_rural_PUMS(
  age_breaks = c(0, 14, 24, 99),
  labels = c("0-13", "14-24", "25+")
)
}
\arguments{
\item{age_breaks}{min age, max age, and any breaks in between (numerical)}

\item{labels}{vector of strings showing age ranges e.g. c("14-24", "25+")}
}
\value{
tibble of length 2*(number of age categories) with four columns
}
\description{
Urban/Rural Breakdown by age group
}
\details{
Pulls numbers and percentages of age groups by urban/rural.
}
\author{
Daniel James Liden
}
