% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninsurance.R
\name{get_uninsurance_PUMS}
\alias{get_uninsurance_PUMS}
\title{Parse PUMS uninsurance data}
\usage{
get_uninsurance_PUMS(
  age_breaks = c(14, 18, 24, 99),
  labels = c("14-18", "19-24", "25+")
)
}
\arguments{
\item{age_breaks}{vector of ages passed to cut(age, age_breaks, include.lowest = TRUE, right=TRUE)}

\item{labels}{vector of labels for age ranges specifed in age breaks}
}
\value{
data frame with percent uninsured by age group and urban/rural
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel James Liden
}
