% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{gradrate}
\alias{gradrate}
\title{2015-2019 4-year graduation rates in Nevada}
\format{
a tibble with 40 rows and 3 columns:
\describe{
\item{Group:}{Race/Ethnicity}
\item{rate:}{4-year graduation rate}
\item{year:}{Graduating class year (e.g. 2015 = class of 2015}
}
}
\source{
\url{https://www.samhsa.gov/data/report/2017-2018-nsduh-state-specific-tables}
}
\usage{
gradrate
}
\description{
A dataset showing the 4-year graduation rate by race in Nevada
}
\keyword{datasets}
