% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_utils.R
\name{load_cols}
\alias{load_cols}
\title{Color Palette utility for plotting}
\usage{
load_cols()
}
\value{
color palette
}
\description{
Loads color palette
}
\details{
Loads color palette for use in plotting functions. Specifically
uses a selection of colors from the pals::cols25() function.
}
\author{
Daniel James Liden
}
