% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{plot_age}
\alias{plot_age}
\title{}
\usage{
plot_age(data = get_agedist_ACS(), path = "~/images/")
}
\arguments{
\item{data}{}

\item{path}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel James Liden
}
