% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu.R
\name{plot_ed_attainment}
\alias{plot_ed_attainment}
\title{}
\usage{
plot_ed_attainment(data = ed_attainment, save = FALSE, path = "~/images/")
}
\arguments{
\item{data}{}

\item{save}{}

\item{path}{}
}
\value{

}
\description{
Educational Attainment in Nevada by Age Group
}
\details{
Comparison of Educational Attainment Levels in Nevada for Nevadans
aged 18-24 and 25+
}
\author{
Daniel James Liden
}
