% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu.R
\name{plot_grad_rate}
\alias{plot_grad_rate}
\title{}
\usage{
plot_grad_rate(data = gradrate, save = FALSE, path = "~/images/")
}
\arguments{
\item{data}{}

\item{save}{}

\item{path}{}
}
\value{

}
\description{
4-year graduation rate in Nevada
}
\details{
Plots the four-year graduation rate in Nevada from 2015 through 2019
by race/ethnicity.
}
\author{
Daniel James Liden
}
