% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty.R
\name{plot_poverty_PUMS_urban_rural}
\alias{plot_poverty_PUMS_urban_rural}
\title{}
\usage{
plot_poverty_PUMS_urban_rural(
  data = get_poverty_PUMS(),
  save = FALSE,
  path = "~/images/"
)
}
\arguments{
\item{data}{}

\item{save}{}

\item{path}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel James Liden
}
