% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suicide.R
\name{plot_suicide_time}
\alias{plot_suicide_time}
\title{Plot suicide rates from 2010 through 2018}
\usage{
plot_suicide_time(data = suicide, save = FALSE, path = "~/images/")
}
\arguments{
\item{data}{}

\item{save}{Whether to save the image; defaults to FALSE}

\item{path}{directory to which to save the image}
}
\description{
Plot suicide rates
}
\details{
Plot suicide rates for NV and the US from 2010 through 2018
for 15-24 and all ages age groups
}
