% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment.R
\name{plot_unemployment_age}
\alias{plot_unemployment_age}
\title{}
\usage{
plot_unemployment_age(
  data = unemployment_age,
  save = FALSE,
  path = "~/images/"
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel James Liden
}
