% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_utils.R
\name{save_fig}
\alias{save_fig}
\title{saves figures with and without GC logo}
\usage{
save_fig(
  fig,
  width = 6.9,
  height = 5,
  name = paste0("fig_", Sys.Date(), ".png"),
  loc = "../figs/"
)
}
\arguments{
\item{fig}{ggplot figure}

\item{width}{default 6.9 in}

\item{height}{default 5 in}

\item{name}{defaults to fig_(date).png. Appended to \code{loc}.}

\item{loc}{directory in which to save figures; default "../figs/"}
}
\value{

}
\description{
Save ggplot figures with some standard formatting
}
\details{
Save ggplot figures with defaults set according to project specifications
and append the guinn center logo to the figures.
}
\author{
Daniel James Liden
}
