% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{suicide}
\alias{suicide}
\title{youth and all-ages suicide rates in NV and nationally}
\format{
a tibble with 36 rows and 6 columns:
\describe{
\item{state}{Nevada or United States}
\item{age_range}{15-24 or all ages}
\item{year}{2010-2018}
\item{deaths}{number of deaths by suicide}
\item{population}{total individuals in age range}
\item{rate_per_100000}{suicides per 100000 people in population}
}
}
\source{
\url{https://wonder.cdc.gov/mcd-icd10.html}
}
\usage{
suicide
}
\description{
A dataset containing suicide rates from 2010 through
2018 for youths aged 15-24 and for all ages in Nevada
and in the United States as a whole.
}
\keyword{datasets}
