% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{unemployment_age}
\alias{unemployment_age}
\title{2019 Unemployment by Age Group in Nevada}
\format{
a tibble with 4 rows and 2 columns:
\describe{
\item{group:}{Age group (total, 16-24, 25-54, 55+}
\item{unemployment rate:}{unemployment rate as \% of labor force}
}
}
\source{
\url{https://www.samhsa.gov/data/report/2017-2018-nsduh-state-specific-tables}
}
\usage{
unemployment_age
}
\description{
A dataset showing the unemployment rate, by age group, in Nevada in 2019.
(Annual Average unemployment rate)
}
\keyword{datasets}
